Fix windows error attempt 2

+++ glyCompare.py
@@ -148,12 +148,12 @@ def structure(args):
     keywords_dict = pipeline_functions.load_para_keywords(project_name, working_addr, reference_addr = reference_addr)
     pipeline_functions.check_init_dir(keywords_dict)
     # Create temperary source data. Delete after GlyCompare is done.
-    try:
+    if os.getcwd().count("/") >= os.getcwd().count("\"):
         os.popen("rm " + "\ ".join(output_path.split(" ")) + "/source_data/" + "\ ".join(abd_path.split(" ")).split("/")[-1])
         os.popen("cp " + "\ ".join(abd_path.split(" ")) + " " + "\ ".join(output_path.split(" ")) + "/source_data/" + "\ ".join(abd_path.split(" ")).split("/")[-1])
         os.popen("rm " + "\ ".join(output_path.split(" ")) + "/source_data/" + "\ ".join(var_path.split(" ")).split("/")[-1])
         os.popen("cp " + "\ ".join(var_path.split(" ")) + " " + "\ ".join(output_path.split(" ")) + "/source_data/" + "\ ".join(var_path.split(" ")).split("/")[-1])
-    except:
+    else:
         os.popen("del \"" + output_path + "\\source_data\\" + PureWindowsPath(abd_path).name + "\"")
         os.popen("copy \"" + abd_path + "\" \"" + output_path + "\\source_data\\" + PureWindowsPath(abd_path).name + "\"")
         os.popen("del \"" + output_path + "\\source_data\\" + PureWindowsPath(var_path).name + "\"")
